requirements.txt
requests
pyyaml
elasticsearch
elasticsearch_dsl
transformers[torch]
scikit-learn
kss
hdbscan
umap-learn
